% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_objects.R
\name{AuditData}
\alias{AuditData}
\title{AuditData Object}
\usage{
AuditData(policyDelta = NULL)
}
\arguments{
\item{policyDelta}{Policy delta between the original policy and the newly set policy}
}
\value{
AuditData object
}
\description{
AuditData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Audit log information specific to Cloud IAM. This message is serializedas an `Any` type in the `ServiceData` message of an`AuditLog` message.
}
