% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_objects.R
\name{Binding}
\alias{Binding}
\title{Google Identity and Access Management (IAM) API Objects 
Manages identity and access control for Google Cloud Platform resources, including the creation of service accounts, which you can use to authenticate to Google and make API calls.}
\usage{
Binding(members = NULL, role = NULL)
}
\arguments{
\item{members}{Specifies the identities requesting access for a Cloud Platform resource}

\item{role}{Role that is assigned to `members`}
}
\value{
Binding object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:59:06
filename: /Users/mark/dev/R/autoGoogleAPI/googleiamv1.auto/R/iam_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Binding Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Associates `members` with a `role`.
}
