% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_objects.R
\name{BindingDelta}
\alias{BindingDelta}
\title{BindingDelta Object}
\usage{
BindingDelta(role = NULL, action = NULL, member = NULL)
}
\arguments{
\item{role}{Role that is assigned to `members`}

\item{action}{The action that was performed on a Binding}

\item{member}{A single identity requesting access for a Cloud Platform resource}
}
\value{
BindingDelta object
}
\description{
BindingDelta Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
One delta entry for Binding. Each individual change (only one member in eachentry) to a binding will be a separate entry.
}
