% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_objects.R
\name{CreateServiceAccountKeyRequest}
\alias{CreateServiceAccountKeyRequest}
\title{CreateServiceAccountKeyRequest Object}
\usage{
CreateServiceAccountKeyRequest(keyAlgorithm = NULL, privateKeyType = NULL)
}
\arguments{
\item{keyAlgorithm}{Which type of key and algorithm to use for the key}

\item{privateKeyType}{The output format of the private key}
}
\value{
CreateServiceAccountKeyRequest object
}
\description{
CreateServiceAccountKeyRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The service account key create request.
}
