% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_objects.R
\name{CreateServiceAccountRequest}
\alias{CreateServiceAccountRequest}
\title{CreateServiceAccountRequest Object}
\usage{
CreateServiceAccountRequest(accountId = NULL, serviceAccount = NULL)
}
\arguments{
\item{accountId}{Required}

\item{serviceAccount}{The ServiceAccount resource to create}
}
\value{
CreateServiceAccountRequest object
}
\description{
CreateServiceAccountRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The service account create request.
}
\seealso{
Other CreateServiceAccountRequest functions: \code{\link{projects.serviceAccounts.create}}
}

