% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_objects.R
\name{DataAccessOptions}
\alias{DataAccessOptions}
\title{DataAccessOptions Object}
\usage{
DataAccessOptions()
}
\value{
DataAccessOptions object
}
\description{
DataAccessOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Write a Data Access (Gin) log
}

