% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_functions.R
\docType{package}
\name{iam_googleAuthR}
\alias{iam_googleAuthR}
\alias{iam_googleAuthR-package}
\title{Google Identity and Access Management API
Manages identity and access control for Google Cloud Platform resources, including the creation of service accounts, which you can use to authenticate to Google and make API calls.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:20:01
filename: /Users/mark/dev/R/autoGoogleAPI/googleiamv1.auto/R/iam_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}
}

