% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_objects.R
\name{ListServiceAccountKeysResponse}
\alias{ListServiceAccountKeysResponse}
\title{ListServiceAccountKeysResponse Object}
\usage{
ListServiceAccountKeysResponse(keys = NULL)
}
\arguments{
\item{keys}{The public keys for the service account}
}
\value{
ListServiceAccountKeysResponse object
}
\description{
ListServiceAccountKeysResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The service account keys list response.
}
