% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_objects.R
\name{ListServiceAccountsResponse}
\alias{ListServiceAccountsResponse}
\title{ListServiceAccountsResponse Object}
\usage{
ListServiceAccountsResponse(nextPageToken = NULL, accounts = NULL)
}
\arguments{
\item{nextPageToken}{To retrieve the next page of results, set}

\item{accounts}{The list of matching service accounts}
}
\value{
ListServiceAccountsResponse object
}
\description{
ListServiceAccountsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The service account list response.
}
