% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_objects.R
\name{LogConfig}
\alias{LogConfig}
\title{LogConfig Object}
\usage{
LogConfig(counter = NULL, dataAccess = NULL, cloudAudit = NULL)
}
\arguments{
\item{counter}{Counter options}

\item{dataAccess}{Data access options}

\item{cloudAudit}{Cloud audit options}
}
\value{
LogConfig object
}
\description{
LogConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specifies what kind of log the caller must write Increment a streamz counter with the specified metric and field names. Metric names should start with a '/', generally be lowercase-only, and end in '_count'. Field names should not contain an initial slash. The actual exported metric names will have '/iam/policy' prepended. Field names correspond to IAM request parameters and field values are their respective values. At present only 'iam_principal', corresponding to IAMContext.principal, is supported. Examples: counter { metric: '/debug_access_count' field: 'iam_principal' } ==> increment counter /iam/policy/backend_debug_access_count {iam_principal=[value of IAMContext.principal]} At this time we do not support: * multiple field names (though this may be supported in the future) * decrementing the counter * incrementing it by anything other than 1
}

