% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_functions.R
\name{projects.serviceAccounts.create}
\alias{projects.serviceAccounts.create}
\title{Creates a ServiceAccountand returns it.}
\usage{
projects.serviceAccounts.create(CreateServiceAccountRequest, name)
}
\arguments{
\item{CreateServiceAccountRequest}{The \link{CreateServiceAccountRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/iam/}{Google Documentation}

Other CreateServiceAccountRequest functions: \code{\link{CreateServiceAccountRequest}}
}
