% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_functions.R
\name{projects.serviceAccounts.list}
\alias{projects.serviceAccounts.list}
\title{Lists ServiceAccounts for a project.}
\usage{
projects.serviceAccounts.list(name, pageToken = NULL, pageSize = NULL)
}
\arguments{
\item{name}{Required}

\item{pageToken}{Optional pagination token returned in an earlier}

\item{pageSize}{Optional limit on the number of service accounts to include in the}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/iam/}{Google Documentation}
}
