% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_functions.R
\name{projects.serviceAccounts.signJwt}
\alias{projects.serviceAccounts.signJwt}
\title{Signs a JWT using a service account's system-managed private key.If no expiry time (`exp`) is provided in the `SignJwtRequest`, IAM sets anan expiry time of one hour by default. If you request an expiry time ofmore than one hour, the request will fail.}
\usage{
projects.serviceAccounts.signJwt(SignJwtRequest, name)
}
\arguments{
\item{SignJwtRequest}{The \link{SignJwtRequest} object to pass to this method}

\item{name}{The resource name of the service account in the following format:}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/iam/}{Google Documentation}

Other SignJwtRequest functions: \code{\link{SignJwtRequest}}
}
