% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_functions.R
\name{projects.serviceAccounts.testIamPermissions}
\alias{projects.serviceAccounts.testIamPermissions}
\title{Tests the specified permissions against the IAM access control policy for the specified IAM resource.}
\usage{
projects.serviceAccounts.testIamPermissions(TestIamPermissionsRequest, resource)
}
\arguments{
\item{TestIamPermissionsRequest}{The \link{TestIamPermissionsRequest} object to pass to this method}

\item{resource}{REQUIRED: The resource for which the policy detail is being requested}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/iam/}{Google Documentation}

Other TestIamPermissionsRequest functions: \code{\link{TestIamPermissionsRequest}}
}

