% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_functions.R
\name{projects.serviceAccounts.update}
\alias{projects.serviceAccounts.update}
\title{Updates a service account. Currently, only the following fields are updatable: 'display_name' . The 'etag' is mandatory.}
\usage{
projects.serviceAccounts.update(ServiceAccount, name)
}
\arguments{
\item{ServiceAccount}{The \link{ServiceAccount} object to pass to this method}

\item{name}{The resource name of the service account in the format 'projects/{project}/serviceAccounts/{account}'}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/iam/}{Google Documentation}

Other ServiceAccount functions: \code{\link{ServiceAccount}}
}

