% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_objects.R
\name{QueryGrantableRolesRequest}
\alias{QueryGrantableRolesRequest}
\title{QueryGrantableRolesRequest Object}
\usage{
QueryGrantableRolesRequest(fullResourceName = NULL)
}
\arguments{
\item{fullResourceName}{Required}
}
\value{
QueryGrantableRolesRequest object
}
\description{
QueryGrantableRolesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The grantable role query request.
}
\seealso{
Other QueryGrantableRolesRequest functions: \code{\link{roles.queryGrantableRoles}}
}
