% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_objects.R
\name{QueryGrantableRolesResponse}
\alias{QueryGrantableRolesResponse}
\title{QueryGrantableRolesResponse Object}
\usage{
QueryGrantableRolesResponse(roles = NULL)
}
\arguments{
\item{roles}{The list of matching roles}
}
\value{
QueryGrantableRolesResponse object
}
\description{
QueryGrantableRolesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The grantable role query response.
}
