% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_objects.R
\name{Role}
\alias{Role}
\title{Role Object}
\usage{
Role(name = NULL, description = NULL, title = NULL)
}
\arguments{
\item{name}{The name of the role}

\item{description}{Optional}

\item{title}{Optional}
}
\value{
Role object
}
\description{
Role Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A role in the Identity and Access Management API.
}
