% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_functions.R
\name{roles.queryGrantableRoles}
\alias{roles.queryGrantableRoles}
\title{Queries roles that can be granted on a particular resource.A role is grantable if it can be used as the role in a binding for a policyfor that resource.}
\usage{
roles.queryGrantableRoles(QueryGrantableRolesRequest)
}
\arguments{
\item{QueryGrantableRolesRequest}{The \link{QueryGrantableRolesRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/iam/}{Google Documentation}

Other QueryGrantableRolesRequest functions: \code{\link{QueryGrantableRolesRequest}}
}
