% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_objects.R
\name{Rule}
\alias{Rule}
\title{Rule Object}
\usage{
Rule(description = NULL, permissions = NULL, action = NULL, in. = NULL,
  notIn = NULL, conditions = NULL, logConfig = NULL)
}
\arguments{
\item{description}{Human-readable description of the rule}

\item{permissions}{A permission is a string of form '}

\item{action}{Required}

\item{in.}{The rule matches if the PRINCIPAL/AUTHORITY_SELECTOR is in this set of entries}

\item{notIn}{The rule matches if the PRINCIPAL/AUTHORITY_SELECTOR is not in this set of entries}

\item{conditions}{Additional restrictions that must be met}

\item{logConfig}{The config returned to callers of tech}
}
\value{
Rule object
}
\description{
Rule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A rule to be applied in a Policy.
}

