% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_objects.R
\name{ServiceAccount}
\alias{ServiceAccount}
\title{ServiceAccount Object}
\usage{
ServiceAccount(name = NULL, projectId = NULL, uniqueId = NULL,
  email = NULL, displayName = NULL, etag = NULL, oauth2ClientId = NULL)
}
\arguments{
\item{name}{The resource name of the service account in the format 'projects/{project}/serviceAccounts/{account}'}

\item{projectId}{@OutputOnly The id of the project that owns the service account}

\item{uniqueId}{@OutputOnly unique and stable id of the service account}

\item{email}{@OutputOnly Email address of the service account}

\item{displayName}{Optional}

\item{etag}{Used to perform a consistent read-modify-write}

\item{oauth2ClientId}{@OutputOnly}
}
\value{
ServiceAccount object
}
\description{
ServiceAccount Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A service account in the Identity and Access Management API. To create a service account, you specify the project_id and account_id for the account. The account_id is unique within the project, and used to generate the service account email address and a stable unique id. All other methods can identify accounts using the format 'projects/{project}/serviceAccounts/{account}'. Using '-' as a wildcard for the project, will infer the project from the account. The account value can be the email address or the unique_id of the service account.
}
\seealso{
Other ServiceAccount functions: \code{\link{projects.serviceAccounts.update}}
}

