% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_objects.R
\name{ServiceAccountKey}
\alias{ServiceAccountKey}
\title{ServiceAccountKey Object}
\usage{
ServiceAccountKey(privateKeyData = NULL, publicKeyData = NULL,
  name = NULL, validBeforeTime = NULL, keyAlgorithm = NULL,
  privateKeyType = NULL, validAfterTime = NULL)
}
\arguments{
\item{privateKeyData}{The private key data}

\item{publicKeyData}{The public key data}

\item{name}{The resource name of the service account key in the following format}

\item{validBeforeTime}{The key can be used before this timestamp}

\item{keyAlgorithm}{Specifies the algorithm (and possibly key size) for the key}

\item{privateKeyType}{The output format for the private key}

\item{validAfterTime}{The key can be used after this timestamp}
}
\value{
ServiceAccountKey object
}
\description{
ServiceAccountKey Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a service account key.A service account has two sets of key-pairs: user-managed, andsystem-managed.User-managed key-pairs can be created and deleted by users.  Users areresponsible for rotating these keys periodically to ensure security oftheir service accounts.  Users retain the private key of these key-pairs,and Google retains ONLY the public key.System-managed key-pairs are managed automatically by Google, and rotateddaily without user intervention.  The private key never leaves Google'sservers to maximize security.Public keys for all service accounts are also published at the OAuth2Service Account API.
}
