% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_objects.R
\name{SignBlobRequest}
\alias{SignBlobRequest}
\title{SignBlobRequest Object}
\usage{
SignBlobRequest(bytesToSign = NULL)
}
\arguments{
\item{bytesToSign}{The bytes to sign}
}
\value{
SignBlobRequest object
}
\description{
SignBlobRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The service account sign blob request.
}
\seealso{
Other SignBlobRequest functions: \code{\link{projects.serviceAccounts.signBlob}}
}

