% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_objects.R
\name{SignBlobResponse}
\alias{SignBlobResponse}
\title{SignBlobResponse Object}
\usage{
SignBlobResponse(keyId = NULL, signature = NULL)
}
\arguments{
\item{keyId}{The id of the key used to sign the blob}

\item{signature}{The signed blob}
}
\value{
SignBlobResponse object
}
\description{
SignBlobResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The service account sign blob response.
}

