% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_objects.R
\name{SignJwtRequest}
\alias{SignJwtRequest}
\title{SignJwtRequest Object}
\usage{
SignJwtRequest(payload = NULL)
}
\arguments{
\item{payload}{The JWT payload to sign, a JSON JWT Claim set}
}
\value{
SignJwtRequest object
}
\description{
SignJwtRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The service account sign JWT request.
}
\seealso{
Other SignJwtRequest functions: \code{\link{projects.serviceAccounts.signJwt}}
}
