% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_objects.R
\name{SignJwtResponse}
\alias{SignJwtResponse}
\title{SignJwtResponse Object}
\usage{
SignJwtResponse(keyId = NULL, signedJwt = NULL)
}
\arguments{
\item{keyId}{The id of the key used to sign the JWT}

\item{signedJwt}{The signed JWT}
}
\value{
SignJwtResponse object
}
\description{
SignJwtResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The service account sign JWT response.
}
