% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_objects.R
\name{CreateServiceAccountKeyRequest}
\alias{CreateServiceAccountKeyRequest}
\title{CreateServiceAccountKeyRequest Object}
\usage{
CreateServiceAccountKeyRequest(privateKeyType = NULL)
}
\arguments{
\item{privateKeyType}{The type of the key requested}
}
\value{
CreateServiceAccountKeyRequest object
}
\description{
CreateServiceAccountKeyRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The service account key create request.
}

