% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_objects.R
\name{ListServiceAccountsResponse}
\alias{ListServiceAccountsResponse}
\title{Google Identity and Access Management API Objects 
Manages identity and access control for Google Cloud Platform resources, including the creation of service accounts, which you can use to authenticate to Google and make API calls.}
\usage{
ListServiceAccountsResponse(accounts = NULL, nextPageToken = NULL)
}
\arguments{
\item{accounts}{The list of matching service accounts}

\item{nextPageToken}{To retrieve the next page of results, set [ListServiceAccountsRequest}
}
\value{
ListServiceAccountsResponse object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:20:01
filename: /Users/mark/dev/R/autoGoogleAPI/googleiamv1.auto/R/iam_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ListServiceAccountsResponse Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The service account list response.
}

