% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_objects.R
\name{ServiceAccountKey}
\alias{ServiceAccountKey}
\title{ServiceAccountKey Object}
\usage{
ServiceAccountKey(name = NULL, privateKeyType = NULL,
  privateKeyData = NULL, validAfterTime = NULL, validBeforeTime = NULL)
}
\arguments{
\item{name}{The resource name of the service account key in the format 'projects/{project}/serviceAccounts/{email}/keys/{key}'}

\item{privateKeyType}{The type of the private key}

\item{privateKeyData}{The key data}

\item{validAfterTime}{The key can be used after this timestamp}

\item{validBeforeTime}{The key can be used before this timestamp}
}
\value{
ServiceAccountKey object
}
\description{
ServiceAccountKey Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a service account key. A service account can have 0 or more key pairs. The private keys for these are not stored by Google. ServiceAccountKeys are immutable.
}

